Create  A BASIC TO-DO WEBAPP
level2 
task 3
Develop a basic to-do app that enables a user to add daily tasks.

The to-do web app must allow users to add new items to a list, and items added to a list must be visible. Also, a user can mark a task “complete” upon its successful completion, and these completed tasks must appear on a separate list, say Completed Tasks.

You can also add a separate section titled Pending Tasks that will list all the tasks that require completion. In such a case, the to-do list will be a union of the Pending Tasks list and the Completed Tasks list. Irrespective of the list, a user must be able to delete or edit tasks from it.

To further increase the web app's complexity or make it more detailed, you can add the date and time when a task is added and/or completed